server/pyproject.toml
[tool.poetry]
name = "friends"
version = "0.1.0"
description = "Project for intership in vk"
authors = ["ZZaRRyaDD <example@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10.5"
Werkzeug = "2.2.2"
watchfiles = "0.16.1"
Pillow = "9.3.0"
django = "3.2.16"
django-model-utils = "4.3.1"
djangorestframework = "3.13.1"
django-cors-headers = "3.13.0"
psycopg2-binary = "2.9.5"
django-extensions = "3.2.1"
dj-database-url = "1.0.0"
djangorestframework-simplejwt = "4.8.0"

[tool.poetry.group.dev.dependencies]
django-stubs = "1.13.0"
pytest = "7.2.0"
pytest-sugar = "0.9.6"
pytest-django = "4.5.2"
djangorestframework-stubs = "1.8.0"
flake8 = "4.0.1"
flake8-isort = "4.2.0"
factory-boy = "3.2.1"
django-debug-toolbar = "3.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
